+++ test/poodinis/containertest.d
@@ -224,6 +224,21 @@ version(unittest) {
 		this(Ola ola) {}
 	}
 
+	struct Thing {
+		int x;
+	}
+
+	class MyConfig {
+		@Value("conf.stuffs")
+		int stuffs;
+
+		@Value("conf.name")
+		string name;
+
+		@Value("conf.thing")
+		Thing thing;
+	}
+
 	// Test register concrete type
 	unittest {
 		auto container = new shared DependencyContainer();
@@ -746,5 +761,41 @@ version(unittest) {
 		container.register!Ola;
 		container.register!Hello;
 		container.resolve!Hello;
+	}
+
+	// Test injection of values
+	unittest {
+		auto container = new shared DependencyContainer();
+		container.register!MyConfig;
+
+		class IntInjector : ValueInjector!int {
+			public override int get(string key) {
+				assert(key == "conf.stuffs");
+				return 364;
+			}
+		}
+
+		class StringInjector : ValueInjector!string {
+			public override string get(string key) {
+				assert(key == "conf.name");
+				return "Le Chef";
+			}
+		}
+
+		class ThingInjector : ValueInjector!Thing {
+			public override Thing get(string key) {
+				assert(key == "conf.thing");
+				return Thing(8899);
+			}
+		}
+
+		container.register!(ValueInjector!int, IntInjector);
+		container.register!(ValueInjector!string, StringInjector);
+		container.register!(ValueInjector!Thing, ThingInjector);
+
+		auto instance = container.resolve!MyConfig;
+		assert(instance.stuffs == 364);
+		assert(instance.name == "Le Chef");
+		assert(instance.thing.x == 8899);
 	}
 }